refactor: Move height and cursor back to model
