+++ Mail/Views/New Message/ComposeMessageView.swift
@@ -84,7 +84,7 @@ struct ComposeMessageView: View {
     @State private var currentSignature: Signature?
     @State private var initialAttachments = [Attachable]()
 
-    @State private var editorModel = RichTextEditorModel()
+    @State private var editorModel = EditorModel()
     @Weak private var scrollView: UIScrollView?
 
     @StateObject private var attachmentsManager: AttachmentsManager
@@ -189,16 +189,17 @@ struct ComposeMessageView: View {
             self.scrollView = scrollView
             scrollView.keyboardDismissMode = .interactive
         }
-        .onChange(of: editorModel.height) { _ in
-            guard let scrollView else { return }
-
-            let fullSize = scrollView.contentSize.height
-            let realPosition = (fullSize - editorModel.height) + editorModel.cursorPosition
-
-            guard realPosition >= 0 else { return }
-            let rect = CGRect(x: 0, y: realPosition, width: 1, height: 1)
-            scrollView.scrollRectToVisible(rect, animated: true)
-        }
+//        .onChange(of: editorModel.height) { _ in
+//            guard let scrollView else { return }
+//
+//            let fullSize = scrollView.contentSize.height
+//            let realPosition = (fullSize - editorModel.height) + editorModel.cursorPosition
+//
+//            guard realPosition >= 0 else { return }
+//            let rect = CGRect(x: 0, y: realPosition, width: 1, height: 1)
+//            scrollView.scrollRectToVisible(rect, animated: true)
+//        }
+        .onChange(of: editorModel.cursorPosition, perform: keepCursorVisible)
         .onChange(of: autocompletionType) { newValue in
             guard newValue != nil else { return }
 
@@ -347,6 +348,12 @@ struct ComposeMessageView: View {
             }
         }
         dismissMessageView()
+    }
+
+    private func keepCursorVisible(_ cursorPosition: CGPoint?) {
+        guard let scrollView, let cursorPosition else {
+            return
+        }
     }
 }
